some updates :eyes:

+++ src/users/datasource/usersDataBase.ts
@@ -5,7 +5,7 @@ import { generateHash } from '../../utils';
 import { User, CreateUserInput, UpdateUserInput } from '../../__generated__/types';
 
 interface UserDocument {
-  _id: string | ObjectId;
+  _id: ObjectId;
   name: string;
   image: string;
   bio: string;
@@ -18,35 +18,34 @@ interface UserDocument {
 }
 
 class Users extends MongoDataSource<UserDocument> {
-  async getAllUsers(): Promise<User[]> {
+  async getAllUsers() {
     const usersCursor = this.collection.find();
-    const usersArray = await usersCursor.toArray() as unknown as User[] ;
+    const usersArray = await usersCursor.toArray();
     return usersArray;
   }
 
-  async getUserById(userId: string): Promise<User> {
-    const user = await this.findOneById(userId) as unknown as User;
+  async getUserById(userId: ObjectId) {
+    const user = await this.findOneById(userId);
     return user;
   }
 
-  async createUser(userInput: CreateUserInput): Promise<User> {
+  async createUser(userInput: CreateUserInput) {
     const hash = await generateHash(userInput.password);
     const createdAt = new Date().toISOString();
     const updatedAt = createdAt;
     const userToInsert = {}
     Object.assign(userToInsert, { ...userInput, createdAt, updatedAt, password: hash })
     const userCreated = await this.collection.insertOne(userToInsert as UserDocument);
-    const user = await this.getUserById(userCreated.insertedId as string);
+    const user = await this.getUserById(userCreated.insertedId);
     return user;
   }
 
-  async updateUser(userId: string, userInput: UpdateUserInput): Promise<User | undefined> {
+  async updateUser(userId: string, userInput: UpdateUserInput) {
     try {
-      const objectId = new ObjectId(userId);
       const fieldsToUpdate = {}
       Object.assign(fieldsToUpdate, userInput);
-      await this.collection.updateOne({ _id: objectId }, { $set: { ...fieldsToUpdate } });
-      return await this.findOneById(objectId) as unknown as User;
+      await this.collection.updateOne({ _id: new ObjectId(userId) }, { $set: { ...fieldsToUpdate } });
+      return await this.findOneById(new ObjectId(userId));
     } catch (error) {
       return undefined;
     }
@@ -54,8 +53,7 @@ class Users extends MongoDataSource<UserDocument> {
 
   async deleteUser(userId: string): Promise<boolean | undefined> {
     try {
-      const _id = new ObjectId(userId);
-      const userDeleted = await this.collection.deleteOne({ _id });
+      const userDeleted = await this.collection.deleteOne({ _id: new ObjectId(userId) });
       return userDeleted.acknowledged;
     } catch (error) {
       return undefined;
